(feat) move wings.model to hummingbot.core.model
