+++ hummingbot/core/model/asset.py
