+++ hummingbot/core/model/market_data.py
