+++ wings/data_source/binance_local_cluster_order_book_data_source.py
@@ -17,8 +17,7 @@ from typing import (
 )
 import conf
 from wings.data_source.local_cluster_order_book_data_source import LocalClusterOrderBookDataSource
-from wings.model.sql_connection_manager import SQLConnectionManager
-from wings.order_book import OrderBook
+from hummingbot.core.model.sql_connection_manager import SQLConnectionManager
 from wings.orderbook.binance_order_book import BinanceOrderBook
