+++ wings/data_source/ddex_local_cluster_order_book_data_source.py
@@ -28,7 +28,7 @@ import conf
 import wings
 from wings.data_source.local_cluster_order_book_data_source import LocalClusterOrderBookDataSource
 from wings.tracker.ddex_active_order_tracker import DDEXActiveOrderTracker
-from wings.model.sql_connection_manager import SQLConnectionManager
+from hummingbot.core.model.sql_connection_manager import SQLConnectionManager
 from wings.order_book_message import (
     OrderBookMessage,
     DDEXOrderBookMessage
