+++ wings/tracker/bittrex_order_book_tracker.py
@@ -3,7 +3,7 @@ import asyncio
 import logging
 from typing import (
     Optional)
-from wings.model.sql_connection_manager import SQLConnectionManager
+from hummingbot.core.model.sql_connection_manager import SQLConnectionManager
 from wings.order_book_tracker import (
     OrderBookTracker,
     OrderBookTrackerDataSourceType
