+++ wings/tracker/ddex_order_book_tracker.py
@@ -13,7 +13,7 @@ from typing import (
     Set)
 from wings.tracker.ddex_active_order_tracker import DDEXActiveOrderTracker
 from wings.data_source.ddex_local_cluster_order_book_data_source import DDEXLocalClusterOrderBookDataSource
-from wings.model.sql_connection_manager import SQLConnectionManager
+from hummingbot.core.model.sql_connection_manager import SQLConnectionManager
 from wings.order_book_tracker import (
     OrderBookTracker,
     OrderBookTrackerDataSourceType
