+++ wings/tracker/huobi_order_book_tracker.py
@@ -4,7 +4,7 @@ import logging
 from typing import (
     Optional)
 import conf
-from wings.model.sql_connection_manager import SQLConnectionManager
+from hummingbot.core.model.sql_connection_manager import SQLConnectionManager
 from wings.order_book_tracker import (
     OrderBookTracker,
     OrderBookTrackerDataSourceType
